feat: Implement NOTAL loop traversal, iterate-stop, and repeat N times
This commit introduces the implementation and testing for three new loop constructs in NOTAL:
- Traversal Loop (i traversal [start..end step N])
- Iterate-Stop Loop (iterate ... stop (condition))
- Repeat N Times (repeat N times ...)

The changes include:
- **Parser:** Updated `src/Parser.cpp` to correctly parse these new loop structures, including handling indentation for loop bodies.
- **Code Generator:** Modified `src/CodeGenerator.cpp` to generate equivalent Pascal code for each loop type. This involved implementing a pre-scan mechanism to collect loop variable declarations and ensuring they are correctly placed in the Pascal `var` block.
- **Test Suite:** Added comprehensive unit tests for each loop construct in `tests/loop_constructs_test.cpp`.
- **Test Infrastructure Refactoring:** Moved common helper functions (`transpile`, `normalizeCode`, `cleanString`) to `tests/test_helpers.h` and `tests/test_helpers.cpp` to reduce code duplication and improve test maintainability.
- **Test Fixes:** Addressed and resolved compiler warnings related to signed/unsigned comparisons in test files (`tests/lexer_new_tokens_test.cpp`, `tests/lexer_test.cpp`, `tests/new_features_test.cpp`, `tests/parser_test.cpp`).
- **Bug Fixes:** Resolved a segmentation fault issue in the code generator's pre-scan logic and corrected an indentation parsing error for `repeat N times` loops.
- **Examples:** Added separate example files for each new loop feature in the `examples/` directory.

All tests pass after these changes, ensuring the new features are correctly implemented and integrated.

+++ tests/lexer_new_tokens_test.cpp
@@ -17,7 +17,7 @@ TEST(LexerNewTokensTest, RecordSyntaxTokens) {
         notal::TokenType::END_OF_FILE
     };
 
-    ASSERT_EQ(tokens.size(), expected_types.size());
+    ASSERT_EQ(static_cast<long long>(tokens.size()), static_cast<long long>(expected_types.size()));
     for (size_t i = 0; i < tokens.size(); ++i) {
         EXPECT_EQ(tokens[i].type, expected_types[i]) << "Token " << i << " has wrong type. Lexeme: " << tokens[i].lexeme;
     }
@@ -38,7 +38,7 @@ TEST(LexerNewTokensTest, EnumSyntaxTokens) {
         notal::TokenType::END_OF_FILE
     };
 
-    ASSERT_EQ(tokens.size(), expected_types.size());
+    ASSERT_EQ(static_cast<long long>(tokens.size()), static_cast<long long>(expected_types.size()));
     for (size_t i = 0; i < tokens.size(); ++i) {
         EXPECT_EQ(tokens[i].type, expected_types[i]) << "Token " << i << " has wrong type. Lexeme: " << tokens[i].lexeme;
     }
@@ -58,7 +58,7 @@ TEST(LexerNewTokensTest, ConstraintSyntaxTokens) {
         notal::TokenType::END_OF_FILE
     };
 
-    ASSERT_EQ(tokens.size(), expected_types.size());
+    ASSERT_EQ(static_cast<long long>(tokens.size()), static_cast<long long>(expected_types.size()));
     for (size_t i = 0; i < tokens.size(); ++i) {
         EXPECT_EQ(tokens[i].type, expected_types[i]) << "Token " << i << " has wrong type. Lexeme: " << tokens[i].lexeme;
     }
