+++ tests/lexer_test.cpp
@@ -26,7 +26,7 @@ ALGORITMA
         notal::TokenType::END_OF_FILE
     };
 
-    ASSERT_EQ(tokens.size(), expected_types.size());
+    ASSERT_EQ(static_cast<long long>(tokens.size()), static_cast<long long>(expected_types.size()));
 
     for (size_t i = 0; i < tokens.size(); ++i) {
         EXPECT_EQ(tokens[i].type, expected_types[i]) << "Token " << i << " has wrong type. Lexeme: " << tokens[i].lexeme;
@@ -55,7 +55,7 @@ TEST(LexerTest, AllOperatorsAndPunctuation) {
         notal::TokenType::ASSIGN, notal::TokenType::END_OF_FILE
     };
 
-    ASSERT_EQ(tokens.size(), expected_types.size());
+    ASSERT_EQ(static_cast<long long>(tokens.size()), static_cast<long long>(expected_types.size()));
 
     for (size_t i = 0; i < tokens.size(); ++i) {
         EXPECT_EQ(tokens[i].type, expected_types[i]) << "Token " << i << " has wrong type. Lexeme: " << tokens[i].lexeme;
